Refactor triangle representation.
* Remove UV-coords from Point class.
* Triangle can be textured and/or have uv-coordinates.
* Efficient static dispatch for both uv-lookup and normal lookup.

+++ include/vcm.h
@@ -97,7 +97,7 @@ private: /* Types: */
             , dVCM {(float) v.dVCM}
             , dVM {(float) v.dVM}
             , continuationPr {(float) v.brdf.continuationPr ()}
-            , length {v.length}
+            , length {(uint8_t) v.length}
         {
             const auto dir = v.brdf.worldDirFix ();
             theta = (float) acos (dir.z);
